moved city cache prefixes
